+++ lib/commands/generateApi/generatePolicies.js
@@ -3,6 +3,7 @@ var path = require('path');
 var mkdirp = require('mkdirp');
 var streamFromArray = require('stream-from-array');
 var assign = require('lodash.assign');
+var serviceUtils = require('../../util/service.js');
 var quota = require('../../policy_templates/quota/quota.js');
 var spike = require('../../policy_templates/spikeArrest/spikeArrest.js');
 var cache = require('../../policy_templates/cache/responseCache.js');
@@ -15,20 +16,6 @@ var verifyApiKey = require('../../policy_templates/security/apikey.js');
 var oauth2 = require('../../policy_templates/security/verifyAccessToken.js');
 var async = require('async');
 
-function servicesToArray (api) {
-  var services = [];
-  if (Array.isArray(api['x-a127-services'])) {
-    services = api['x-a127-services'];
-  } else {
-    Object.keys(api['x-a127-services']).forEach(function (key) {
-      var service = api['x-a127-services'][key];
-      service['name'] = key;
-      services.push(service);
-    });
-  }
-  return services;
-};
-
 module.exports = function generatePolicies(apiProxy, options, api, cb) {
   var destination = options.destination || path.join(__dirname, '../../../api_bundles');
   if(destination.substr(-1) === '/') {
@@ -36,7 +23,7 @@ module.exports = function generatePolicies(apiProxy, options, api, cb) {
   }
   var rootDirectory = destination + "/" + apiProxy + "/apiproxy";
   var validationCount = 0;
-  var services = servicesToArray(api);
+  var services = serviceUtils.servicesToArray(api);
   async.each(services, function(service, callback) {
     // Perform operation on file here.
     var serviceName = service.name;
