+++ lib/commands/generateApi/generateProxyEndPoint.js
@@ -2,6 +2,7 @@ var builder = require('xmlbuilder');
 var fs = require('fs');
 var path = require('path');
 var quota = require('../../policy_templates/quota/quota.js');
+var serviceUtils = require('../../util/service.js');
 
 
 module.exports = function generateProxyEndPoint(apiProxy, options, api, cb) {
@@ -20,12 +21,12 @@ module.exports = function generateProxyEndPoint(apiProxy, options, api, cb) {
   var raiseFaultName;
   var requestPipe = preFlow.ele('Request');
   var responsePipe = preFlow.ele('Response');
-  for (var service in api['x-a127-services']) {
-    var serviceItem = api['x-a127-services'][service];
+  var services = serviceUtils.servicesToArray(api);
+  services.forEach(function (serviceItem) {
     if (serviceItem.provider === 'x-cors') {
-      useCors = service;
+      useCors = serviceItem.name;
       var step = responsePipe.ele('Step', {});
-      step.ele('Name', {}, service);
+      step.ele('Name', {}, serviceItem.name);
       step.ele('Condition', {}, 'request.verb != "OPTIONS"');
     }
     if (serviceItem.provider === 'x-input-validation' &&
@@ -64,7 +65,7 @@ module.exports = function generateProxyEndPoint(apiProxy, options, api, cb) {
         step.ele('Name', {}, raiseFaultName);
       }
     }
-  }
+  });
 
   var flows = root.ele('Flows', {});
 
@@ -76,16 +77,6 @@ module.exports = function generateProxyEndPoint(apiProxy, options, api, cb) {
     var requestStep = preFlowResponse.ele('Step');
     requestStep.ele('Name',{}, useCors);
   }
-  // <Flow name="OptionsPreFlight">
-  //     <Request/>
-  //     <Response>
-  //         <Step>
-  //             <Name>Add-CORS</Name>
-  //         </Step>
-  //     </Response>
-  // <Condition>request.verb == "OPTIONS"</Condition>
-  // </Flow>
-
 
   var allowedVerbs = ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS', 'HEAD', 'TRACE', 'CONNECT', 'PATCH'];
   for (var apiPath in api.paths) {
