+++ test/commands/generateApiCorsArray.js
@@ -7,11 +7,11 @@ var generateSkeleton = require('../../lib/commands/generateApi/generateSkeleton.
 var fs = require('fs');
 var xml2js = require('xml2js');
 
-describe('generateApi with CORS proxy', function() {
+describe('generateApi with CORS proxy (array)', function() {
   var options = {
-    source : path.join(__dirname, '/openapi_files/cors.yaml'),
+    source : path.join(__dirname, '/openapi_files/cors-array.yaml'),
     destination : path.join(__dirname, '../../api_bundles'),
-    apiProxy :'petStoreCors'
+    apiProxy :'petStoreCorsArray'
   };
 
   describe('generate', function() {
@@ -91,56 +91,6 @@ describe('generateApi with CORS proxy', function() {
         should.exist(result.TargetEndpoint.PreFlow[0].Request[0], 'Request found in PreFlow');
         should.equal(result.TargetEndpoint.PreFlow[0].Request[0].length, 0, 'Request step not found in PreFlow');
         done();
-      });
-    });
-
-    describe('virtualhosts option', function(done) {
-      it("missing -v flag should generate both default and secure", function(done) {
-        options.apiProxy = 'petStoreVirtualBoth';
-        generateApi.generateApi(options.apiProxy, options, function(err, reply) {
-          should.equal(err, null);
-          var proxiesFilePath = path.join(options.destination, options.apiProxy, '/apiproxy/proxies/default.xml');
-          var proxiesFileData = fs.readFileSync(proxiesFilePath);
-          var parser = new xml2js.Parser();
-          parser.parseString(proxiesFileData, function (err, result) {
-            result.should.have.property('ProxyEndpoint').property('HTTPProxyConnection');
-            var vhost = result.ProxyEndpoint.HTTPProxyConnection[0].VirtualHost;
-            vhost.should.eql(['default','secure'], 'secure virtual host found');
-            done();
-          });
-        })
-      });
-      it("-v 'secure' should generate secure virtual host", function(done) {
-        options.apiProxy = 'petStoreVirtualVirtual';
-        options.virtualhosts = 'secure';
-        generateApi.generateApi(options.apiProxy, options, function(err, reply) {
-          should.equal(err, null);
-          var proxiesFilePath = path.join(options.destination, options.apiProxy, '/apiproxy/proxies/default.xml');
-          var proxiesFileData = fs.readFileSync(proxiesFilePath);
-          var parser = new xml2js.Parser();
-          parser.parseString(proxiesFileData, function (err, result) {
-            result.should.have.property('ProxyEndpoint').property('HTTPProxyConnection');
-            var vhost = result.ProxyEndpoint.HTTPProxyConnection[0].VirtualHost;
-            vhost.should.eql(['secure'], 'secure virtual host found');
-            done();
-          });
-        })
-      });
-      it("-v 'default' should generate default virtual host", function(done) {
-        options.apiProxy = 'petStoreVirtualDefault';
-        options.virtualhosts = 'default';
-        generateApi.generateApi(options.apiProxy, options, function(err, reply) {
-          should.equal(err, null);
-          var proxiesFilePath = path.join(options.destination, options.apiProxy, '/apiproxy/proxies/default.xml');
-          var proxiesFileData = fs.readFileSync(proxiesFilePath);
-          var parser = new xml2js.Parser();
-          parser.parseString(proxiesFileData, function (err, result) {
-            result.should.have.property('ProxyEndpoint').property('HTTPProxyConnection');
-            var vhost = result.ProxyEndpoint.HTTPProxyConnection[0].VirtualHost;
-            vhost.should.eql(['default'], 'secure virtual host found');
-            done();
-          });
-        })
       });
     });
